% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_EM_var.R
\name{loglik_EM_var}
\alias{loglik_EM_var}
\title{Calculate the covariance matrix}
\usage{
loglik_EM_var(Dstar, X, V, L, theta, beta, gamma, expected)
}
\arguments{
\item{Dstar}{A numeric vector containing the observed outcome of interest, should be coded as 0/1.}

\item{X}{A numeric vector containing the predictor of interest.}

\item{V}{A numeric vector of covariate in sensitivity model.}

\item{L}{A numeric vector of covariate in specificity model.}

\item{theta}{The estimates of parameters in general binary regression model.}

\item{beta}{The estimates of coefficients in sensitivity model.}

\item{gamma}{The estimates of coefficients in specificity model.}

\item{expected}{Whether or not to consider the expected information matrix.}
}
\value{
'var', the covariance matrix of the final estimates.
}
\description{
\code{loglik_EM_var} Consider the "Huber Sandwich Estimator" to calculate the covariance matrix using the estimates obtained via expectation-maximization algorithm.
}
