% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_EM.R
\name{loglik_EM}
\alias{loglik_EM}
\title{Estimate parameters using the expectation-maximization algorithm}
\usage{
loglik_EM(Dstar, X, V, L, start, tol = 1e-08, maxit = 1000)
}
\arguments{
\item{Dstar}{A numeric vector containing the observed outcome of interest, should be coded as 0/1.}

\item{X}{A numeric vector containing the predictor of interest.}

\item{V}{A numeric vector of covariate in sensitivity model.}

\item{L}{A numeric vector of covariate in specificity model.}

\item{start}{Numeric vectors of starting values for theta, beta and gamma. Theta is the parameter of
binary regression model. Beta and gamma are regression coefficients in sensitivity and specificity model, respectively.}

\item{tol}{Numerical precision.}

\item{maxit}{Maximum number of interations to solve the estimating equations.}
}
\value{
A list
\describe{
\item{par}{the final estimates of the coeficients organized as (theta, beta and gamma).}
\item{param.seq}{the sequence of estimates at each itertaion.}
\item{loglik.seq}{the sequence of the log likelihood at each itertaion.}
\item{Dstar}{A numeric vector of the input observed outcome.}
\item{X}{A numeric vector of the input predictor variable.}
\item{V}{A numeric vector of the input covariate in sensitivity model.}
\item{L}{A numeric vector of the input covariate in specificity model.}
}
}
\description{
\code{loglik_EM} Estimate the parameters using the expectation-maximization algorithm based on the complete log-likelihood of observed data in the presence of missing data.
}
\examples{
library(DMSS)
library(rje)

# Generate the predictor variable and covariates
X = rnorm(10000)
L = rnorm(10000)
V = rnorm(10000)

# Generate the covariate-related sensitivity and specificity
theta = c(-2,3)
p_Dtrue = expit(cbind(1,X)\%*\% theta)
beta = c(0.45,0.5)
sensitivity = expit(cbind(1,V)\%*\% beta)
gamma = c(4.5,1)
specificity = expit(cbind(1,L)\%*\% gamma)

# Generate the observed outcome variable
p_Dstar = (1-specificity)+(specificity+sensitivity -1)*p_Dtrue
Dstar = rbinom(10000,1,p_Dstar)

# initial value
start = c(-1,1,0,0,3,0)
result = loglik_EM (Dstar,X,V,L,start,tol = 1e-8, maxit = 1000)
}
